Unprogramming old habits
More often than not, I'm logging onto systems that just have the
rudimentary vi installed. And the reminder that vim is not
installed when I type it out is just getting in the way, when all I
really mean is just give me an editor so I can change something.

So here's me forcing myself to always type vi, and alias it to
greater things on systems where I actually care about the editor.