fedora-ws-setup: fix reboot logic?
So if the program exits with 0 then bash interprets it as true,
which is not what the program does. It spits out 1 if we need a
reboot.

Using the inverse [ ! needs-restarting -r ] logic does not read
well, so we're better off checking the return code and doing
something based on that instead.