Install rpmfusion agnostically
So a couple things are happening here.

Mirrors are too unreliable during install, so we stick to the
original source with `--baseurl=`

Unless you use `repo --install ...` the repo is just there during
installation, and not on the installed system. These repos are
rather limited with no debuginfo or source, but by having these
limited repos for rpmfusion available during installation, we can
install the packages `rpmfusion-{free,nonfree}-release` to our
system.

We want the installation to work for both fedora-40 and fedora-41,
so we try to make the agnostic by leveraging %pre, sourcing
/etc/os-release and templating out to a file we can %include from
the kickstart.

And finally renaming this kickstart to reflect that it can
(hopefully) apply to any present or upcoming fedora releases.