setup-fedora-ws: fix reboot check
With `set -euo pipefail` there's no point in checking the return
code of a program this way, because the script exits on non-zero
return codes, thereby not checking the return code nor executing
`systemctl reboot`

I like this new solution with the or-check where the command on the
right-hand side will run only if the command on the left-hand side
fails. (i.e. returns non-zero)