fedora.ks: custom partitioning
This ended up being slightly more complex than it probably needed. Once I added
a the second nvme for /home to lizzie, it dawned on me that I could always keep
/home and just reinstall the OS on the primary nvme as I see fit.

Then I thought what if I'd like to keep home on other systems aswell? And what
if they were bios based? And a whole other bunch of what ifs. And doing stuff
dynamically, and it got stupid fast.

lee is a r720xd with 12 drives in front, 1 of 2 possible in the rear, on a H310
mini mono flashed to IT mode. So the drives keep jumping around.